backend/requirements.txt
Flask
flask-cors
openai
python-dotenv
requests